requirements.txt
# python==3.10.0
joblib==1.2.0
matplotlib==3.7.0
networkx==3.0
numpy>=1.23
scipy==1.10.1
scikit_learn==1.2.1
pandas==1.5.3
plotly==5.13.0
torch==1.13.1
pyro_ppl==1.8.4
PyYAML==6.0
seaborn==0.12.2
tqdm==4.64.1
pynumdiff==0.1.2
gpytorch==1.9.1
aim>=3.16.0